add new designed chart settings to mobile chart settings
