+++ src/pages/Chart/ChartSettings/ChartSettings.tsx
@@ -1,26 +1,15 @@
-import { MouseEvent, useContext, useEffect, useRef, useState } from 'react';
+import { useContext, useEffect, useRef, useState } from 'react';
 import * as d3 from 'd3';
 import {
     ActionButtonContainer,
     ChartSettingsContainer,
-    CheckList,
-    CheckListContainer,
     CloseButton,
-    ColorList,
-    ColorOptions,
-    ColorPickerContainer,
     ContextMenu,
-    ContextMenuContextText,
     ContextMenuFooter,
     ContextMenuHeader,
     ContextMenuHeaderText,
     FooterButtons,
     FooterContextText,
-    Icon,
-    OptionColor,
-    SelectionContainer,
-    StyledCheckbox,
-    StyledSelectbox,
 } from './ChartSettingsCss';
 import { VscClose } from 'react-icons/vsc';
 import {
@@ -28,26 +17,18 @@ import {
     LocalChartSettingsIF,
 } from '../../../contexts/ChartContext';
 
-import { SketchPicker } from 'react-color';
 import { PoolContext } from '../../../contexts/PoolContext';
 import { BrandContext } from '../../../contexts/BrandContext';
 import Spinner from '../../../components/Global/Spinner/Spinner';
-import { TradeDataContext } from '../../../contexts/TradeDataContext';
 import { LS_KEY_CHART_CONTEXT_SETTINGS } from '../../platformAmbient/Chart/ChartUtils/chartConstants';
 import {
     chartItemStates,
     getCssVariable,
 } from '../../platformAmbient/Chart/ChartUtils/chartUtils';
-import {
-    ColorPickerTab,
-    DropDownList,
-    DropDownListContainer,
-    LabelSettingsArrow,
-    ListItem,
-} from '../../platformAmbient/Chart/Draw/FloatingToolbar/FloatingToolbarSettingsCss';
+import ChartSettingsContent from './ChartSettingsContent';
 
 interface ContextMenuIF {
-    contextMenuPlacement: { top: number; left: number; isReversed: boolean };
+    contextMenuPlacement?: { top: number; left: number; isReversed: boolean };
     setContextmenu: React.Dispatch<React.SetStateAction<boolean>>;
     setColorChangeTrigger: React.Dispatch<React.SetStateAction<boolean>>;
     chartItemStates: chartItemStates;
@@ -68,9 +49,10 @@ interface ContextMenuIF {
         React.SetStateAction<boolean>
     >;
     render: () => void;
+    isMobile: boolean;
 }
 
-interface ColorObjIF {
+export interface ColorObjIF {
     selectedColor: string | undefined;
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     replaceSelector: string;
@@ -86,13 +68,14 @@ export default function ChartSettings(props: ContextMenuIF) {
         // setChartThemeColors,
         defaultChartSettings,
         setLocalChartSettings,
-        render,
+        // render,
         setColorChangeTrigger,
         isCondensedModeEnabled,
         setIsCondensedModeEnabled,
         setShouldDisableChartSettings,
         closeOutherChartSetting,
         setCloseOutherChartSetting,
+        isMobile,
     } = props;
 
     const {
@@ -115,12 +98,6 @@ export default function ChartSettings(props: ContextMenuIF) {
     const [applyDefault, setApplyDefault] = useState(false);
     const [reverseColorObj, setReverseColorObj] = useState(false);
 
-    const {
-        baseToken: { symbol: baseTokenSymbol },
-        quoteToken: { symbol: quoteTokenSymbol },
-        isDenomBase,
-    } = useContext(TradeDataContext);
-
     useEffect(() => {
         d3.select(contextMenuRef.current).on(
             'contextmenu',
@@ -133,13 +110,15 @@ export default function ChartSettings(props: ContextMenuIF) {
     useEffect(() => {
         const screenHeight = window.innerHeight;
 
-        const diff = screenHeight - contextMenuPlacement.top;
+        if (!isMobile && contextMenuPlacement) {
+            const diff = screenHeight - contextMenuPlacement.top;
 
-        setReverseColorObj(
-            (contextMenuPlacement.isReversed && diff < 260) ||
-                (!contextMenuPlacement.isReversed && diff < 700),
-        );
-    }, [contextMenuPlacement]);
+            setReverseColorObj(
+                (contextMenuPlacement.isReversed && diff < 260) ||
+                    (!contextMenuPlacement.isReversed && diff < 700),
+            );
+        }
+    }, [contextMenuPlacement, isMobile]);
 
     useEffect(() => {
         if (closeOutherChartSetting) {
@@ -148,46 +127,6 @@ export default function ChartSettings(props: ContextMenuIF) {
             setCloseOutherChartSetting(false);
         }
     }, [closeOutherChartSetting]);
-
-    const handleCandleColorPicker = (
-        replaceSelector: string,
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        newColor: any,
-    ) => {
-        const colorRgbaCode =
-            'rgba(' +
-            newColor.rgb.r +
-            ',' +
-            newColor.rgb.g +
-            ',' +
-            newColor.rgb.b +
-            ',' +
-            newColor.rgb.a +
-            ')';
-
-        const replaceColor = d3.color(colorRgbaCode);
-
-        if (replaceSelector && replaceColor) {
-            chartThemeColors[replaceSelector] = replaceColor;
-
-            setColorChangeTrigger(true);
-
-            setSelectedColorObj((prev: ColorObjIF | undefined) => {
-                if (prev) {
-                    const colorRep = {
-                        selectedColor: replaceColor.toString(),
-                        replaceSelector: prev.replaceSelector,
-                        index: prev.index,
-                        placement: prev.placement,
-                    } as ColorObjIF;
-
-                    return colorRep;
-                }
-            });
-
-            render();
-        }
-    };
 
     const handleApplyDefaults = (
         chartSettings: LocalChartSettingsIF,
@@ -206,33 +145,36 @@ export default function ChartSettings(props: ContextMenuIF) {
         );
 
         const upCandleBodyColor = isDefault
-            ? getCssVariable(skin.active, chartSettings.chartColors.upCandleBodyColor)
+            ? getCssVariable(
+                  skin.active,
+                  chartSettings.chartColors.upCandleBodyColor,
+              )
             : d3.color(chartSettings.chartColors.upCandleBodyColor);
 
         const downCandleBodyColor = isDefault
             ? getCssVariable(
-                skin.active,
+                  skin.active,
                   chartSettings.chartColors.downCandleBodyColor,
               )
             : d3.color(chartSettings.chartColors.downCandleBodyColor);
 
         const selectedDateFillColor = isDefault
             ? getCssVariable(
-                skin.active,
+                  skin.active,
                   chartSettings.chartColors.selectedDateFillColor,
               )
             : d3.color(chartSettings.chartColors.selectedDateFillColor);
 
         const downCandleBorderColor = isDefault
             ? getCssVariable(
-                skin.active,
+                  skin.active,
                   chartSettings.chartColors.downCandleBorderColor,
               )
             : d3.color(chartSettings.chartColors.downCandleBorderColor);
 
         const upCandleBorderColor = isDefault
             ? getCssVariable(
-                skin.active,
+                  skin.active,
                   chartSettings.chartColors.upCandleBorderColor,
               )
             : d3.color(chartSettings.chartColors.upCandleBorderColor);
@@ -247,7 +189,7 @@ export default function ChartSettings(props: ContextMenuIF) {
 
         const selectedDateStrokeColor = isDefault
             ? getCssVariable(
-                skin.active,
+                  skin.active,
                   chartSettings.chartColors.selectedDateStrokeColor,
               )
             : d3.color(chartSettings.chartColors.selectedDateStrokeColor);
@@ -342,91 +284,20 @@ export default function ChartSettings(props: ContextMenuIF) {
         };
     };
 
-    const handlePriceInChange = (option: string) => {
-        setIsTradeDollarizationEnabled(
-            option !== quoteTokenSymbol && option !== baseTokenSymbol,
-        );
-        setPriceInOption(option);
-    };
-
     const [selectedColorObj, setSelectedColorObj] = useState<
         ColorObjIF | undefined
     >(undefined);
 
     const [isSelecboxActive, setIsSelecboxActive] = useState(false);
 
-    const [priceInOption, setPriceInOption] = useState<string>(
-        !isTradeDollarizationEnabled
-            ? isDenomBase
-                ? quoteTokenSymbol
-                : baseTokenSymbol
-            : 'USD',
-    );
-
-    const checkListContent = [
-        {
-            checked: showVolume,
-            action: setShowVolume,
-            selection: 'Show Volume',
-        },
-        {
-            checked: showTvl,
-            action: setShowTvl,
-            selection: 'Show TVL',
-        },
-        {
-            checked: showFeeRate,
-            action: setShowFeeRate,
-            selection: 'Show Fee Rate',
-        },
-        {
-            checked: isCondensedModeEnabled,
-            action: setIsCondensedModeEnabled,
-            selection: 'Hide empty candles',
-        },
-    ];
-
-    const selectionContent = [
-        {
-            selection: 'Show prices in',
-            action: handlePriceInChange,
-            options: [isDenomBase ? quoteTokenSymbol : baseTokenSymbol, 'USD'],
-        },
-    ];
-
-    const colorPickerContent = [
-        {
-            selection: 'Candle Body',
-            actionHandler: 'body',
-            action: handleCandleColorPicker,
-            upColor: 'upCandleBodyColor',
-            downColor: 'downCandleBodyColor',
-            exclude: [''],
-        },
-        {
-            selection: 'Candle Borders',
-            actionHandler: 'border',
-            action: handleCandleColorPicker,
-            upColor: 'upCandleBorderColor',
-            downColor: 'downCandleBorderColor',
-            exclude: [''],
-        },
-        {
-            selection: 'Liquidity Area',
-            actionHandler: 'liq',
-            action: handleCandleColorPicker,
-            upColor: 'liqAskColor',
-            downColor: 'liqBidColor',
-            exclude: ['futa'],
-        },
-    ];
-
     return (
         <ChartSettingsContainer
             ref={contextMenuRef}
-            top={contextMenuPlacement.top}
-            left={contextMenuPlacement.left}
-            isReversed={contextMenuPlacement.isReversed}
+            top={contextMenuPlacement ? contextMenuPlacement.top : 0}
+            left={contextMenuPlacement ? contextMenuPlacement.left : 0}
+            isReversed={
+                contextMenuPlacement ? contextMenuPlacement.isReversed : false
+            }
             onClick={() => {
                 setShouldDisableChartSettings(true);
                 setSelectedColorObj(undefined);
@@ -445,238 +316,22 @@ export default function ChartSettings(props: ContextMenuIF) {
                     </CloseButton>
                 </ContextMenuHeader>
 
-                <CheckListContainer>
-                    {checkListContent.map((item, index) => (
-                        <CheckList key={index}>
-                            <StyledCheckbox
-                                checked={item.checked}
-                                onClick={() => item.action(!item.checked)}
-                            >
-                                <Icon
-                                    viewBox='0 0 24 24'
-                                    style={{ width: '24px', height: '24px' }}
-                                >
-                                    <polyline points='20 6 9 17 4 12' />
-                                </Icon>
-                            </StyledCheckbox>
-                            <ContextMenuContextText>
-                                {['futa'].includes(platformName)
-                                    ? item.selection.toUpperCase()
-                                    : item.selection}
-                            </ContextMenuContextText>
-                        </CheckList>
-                    ))}
-                </CheckListContainer>
-
-                <SelectionContainer>
-                    {selectionContent.map((item, index) => (
-                        <CheckList key={index}>
-                            <ContextMenuContextText>
-                                {['futa'].includes(platformName)
-                                    ? item.selection.toUpperCase()
-                                    : item.selection}
-                            </ContextMenuContextText>
-                            <StyledSelectbox
-                                onClick={(event: MouseEvent<HTMLElement>) => {
-                                    event.stopPropagation();
-                                    setIsSelecboxActive(!isSelecboxActive);
-                                }}
-                            >
-                                <ContextMenuContextText>
-                                    {priceInOption}
-                                </ContextMenuContextText>
-                                <LabelSettingsArrow
-                                    isActive={isSelecboxActive}
-                                    width={8}
-                                    height={8}
-                                ></LabelSettingsArrow>
-                            </StyledSelectbox>
-
-                            {isSelecboxActive && (
-                                <div
-                                    style={{
-                                        position: 'relative',
-                                        left: '-42.5%',
-                                        top: '38%',
-                                    }}
-                                >
-                                    <DropDownListContainer>
-                                        <DropDownList width={90}>
-                                            {item.options.map(
-                                                (option, index) => (
-                                                    <ListItem
-                                                        style={{
-                                                            padding:
-                                                                '1px 10px 1px 10px',
-                                                        }}
-                                                        backgroundColor={
-                                                            priceInOption ===
-                                                            option
-                                                                ? 'var(--accent1)'
-                                                                : 'var(--dark1)'
-                                                        }
-                                                        key={index}
-                                                        onClick={(
-                                                            event: MouseEvent<HTMLElement>,
-                                                        ) => {
-                                                            event.stopPropagation();
-                                                            item.action(option);
-                                                        }}
-                                                    >
-                                                        {option}
-                                                    </ListItem>
-                                                ),
-                                            )}
-                                        </DropDownList>
-                                    </DropDownListContainer>
-                                </div>
-                            )}
-                        </CheckList>
-                    ))}
-                </SelectionContainer>
-
-                <ColorPickerContainer>
-                    {colorPickerContent.map(
-                        (item, index) =>
-                            !item.exclude.includes(platformName) && (
-                                <ColorList key={index}>
-                                    <ContextMenuContextText>
-                                        {['futa'].includes(platformName)
-                                            ? item.selection.toUpperCase()
-                                            : item.selection}
-                                    </ContextMenuContextText>
-
-                                    <ColorOptions>
-                                        <OptionColor
-                                            backgroundColor={chartThemeColors[
-                                                item.upColor
-                                            ]?.toString()}
-                                            onClick={(
-                                                event: MouseEvent<HTMLElement>,
-                                            ) => {
-                                                event.stopPropagation();
-                                                setSelectedColorObj((prev) => {
-                                                    const selectedObj = {
-                                                        selectedColor:
-                                                            chartThemeColors[
-                                                                item.upColor
-                                                            ]?.toString(),
-                                                        replaceSelector:
-                                                            item.upColor,
-                                                        index: index,
-                                                        placement: 'left',
-                                                    };
-
-                                                    const result =
-                                                        prev === undefined ||
-                                                        prev.index !== index ||
-                                                        prev.placement !==
-                                                            'left'
-                                                            ? selectedObj
-                                                            : undefined;
-
-                                                    setShouldDisableChartSettings(
-                                                        result === undefined,
-                                                    );
-
-                                                    return result;
-                                                });
-                                            }}
-                                        ></OptionColor>
-                                        {item.downColor !== '' && (
-                                            <OptionColor
-                                                backgroundColor={chartThemeColors[
-                                                    item.downColor
-                                                ]?.toString()}
-                                                onClick={(
-                                                    event: MouseEvent<HTMLElement>,
-                                                ) => {
-                                                    event.stopPropagation();
-                                                    setSelectedColorObj(
-                                                        (prev) => {
-                                                            const selectedObj =
-                                                                {
-                                                                    selectedColor:
-                                                                        chartThemeColors[
-                                                                            item
-                                                                                .downColor
-                                                                        ]?.toString(),
-                                                                    replaceSelector:
-                                                                        item.downColor,
-                                                                    index: index,
-                                                                    placement:
-                                                                        'right',
-                                                                };
-
-                                                            const result =
-                                                                prev ===
-                                                                    undefined ||
-                                                                prev.index !==
-                                                                    index ||
-                                                                prev.placement !==
-                                                                    'right'
-                                                                    ? selectedObj
-                                                                    : undefined;
-
-                                                            setShouldDisableChartSettings(
-                                                                result ===
-                                                                    undefined,
-                                                            );
-
-                                                            return result;
-                                                        },
-                                                    );
-                                                }}
-                                            ></OptionColor>
-                                        )}
-
-                                        {selectedColorObj &&
-                                            selectedColorObj.index ===
-                                                index && (
-                                                <ColorPickerTab
-                                                    style={{
-                                                        position: 'fixed',
-                                                        zIndex: 199,
-                                                        transform:
-                                                            'translate(' +
-                                                            (selectedColorObj.placement ===
-                                                            'left'
-                                                                ? -90
-                                                                : -60) +
-                                                            'px, ' +
-                                                            (reverseColorObj
-                                                                ? '-325px)'
-                                                                : '10px)'),
-                                                    }}
-                                                    onClick={(
-                                                        event: MouseEvent<HTMLElement>,
-                                                    ) =>
-                                                        event.stopPropagation()
-                                                    }
-                                                >
-                                                    <SketchPicker
-                                                        color={
-                                                            selectedColorObj.selectedColor
-                                                        }
-                                                        width={'170px'}
-                                                        onChange={(
-                                                            color,
-                                                            event,
-                                                        ) => {
-                                                            event.stopPropagation();
-                                                            item.action(
-                                                                selectedColorObj.replaceSelector,
-                                                                color,
-                                                            );
-                                                        }}
-                                                    />
-                                                </ColorPickerTab>
-                                            )}
-                                    </ColorOptions>
-                                </ColorList>
-                            ),
-                    )}
-                </ColorPickerContainer>
+                <ChartSettingsContent
+                    chartThemeColors={chartThemeColors}
+                    isCondensedModeEnabled={isCondensedModeEnabled}
+                    setIsCondensedModeEnabled={setIsCondensedModeEnabled}
+                    setShouldDisableChartSettings={
+                        setShouldDisableChartSettings
+                    }
+                    chartItemStates={props.chartItemStates}
+                    setColorChangeTrigger={setColorChangeTrigger}
+                    isSelecboxActive={isSelecboxActive}
+                    setIsSelecboxActive={setIsSelecboxActive}
+                    selectedColorObj={selectedColorObj}
+                    setSelectedColorObj={setSelectedColorObj}
+                    reverseColorObj={reverseColorObj}
+                    // render={render}
+                />
 
                 <ContextMenuFooter>
                     <FooterButtons
